CloudComponents/ImageAnalyzer/requirements.txt
functions-framework
google-cloud-storage
google-cloud-vision
ImageHash
Pillow 